requirements.txt
requests
langchain
langchain-community
ipykernel
faiss-cpu
chromadb
beautifulsoup4
lxml
langchain-google-genai
google-generativeai
pypdf
python-dotenv